// src/types/auth.service.types.ts
import {EUserRole} from "@/types/user.types.ts";

export type TAuthSigninRequestParam = {
    email: string
    password: string
    role: EUserRole
}

export type TAuthSignupRequestParam = {
    name: string
    email: string
    password: string
    phone?: string
    role: EUserRole
}

export type TAuthResponse = {
    success: boolean
    token: string
    user: {
        id: string
        name: string
        email: string
        phone?: string
        role: EUserRole
    }
    message?: string
}